requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asn1crypto==1.5.1
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
filelock==3.16.1
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
imbalanced-learn==0.13.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
lightgbm==4.5.0
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
platformdirs==4.3.6
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.15.0
six==1.17.0
sklearn-compat==0.1.3
slack_sdk==3.34.0
slackclient==2.9.4
snowflake-connector-python==3.12.4
sortedcontainers==2.4.0
threadpoolctl==3.5.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
